test(assisstant): add unit tests for commands, queries, and validation
Validated CreateAssisstantCommand and mapping logic with DTO refinements

Covered GetAll and GetById queries with focused assertions

Tested validators for input integrity and edge case handling

+++ src/Application/Assisstants/Mapper/AssisstantMapper.cs
@@ -11,7 +11,7 @@ namespace Application.Assisstants.Mapper
         {
             return new Assisstant
             {
-               AssignedBy = dto.AssignedBy,
+               AssignedBy = (int)dto.AssignedBy,
              //  UserId = dto.UserId,
